Add CheckRetry callback to the Client
There currently is no way to set a policy to determine what
constitutes a retry condition. We can't stop retries on an error we know
is unrecoverable, and we can't continue retrying on an error that we
know is temporary. For example, we probably want to backoff and retry on
429, because that's what it's for, though some systems may send a
different code and/or message, and may even provide a retry time.

This adds a CheckRestry callback to the client that will be called after
each request. If CheckRetry returns false, the client will
immediately return the response. If CheckRetry returns an error with
false, that error will be returned instead of the original request
error.

Add a check for an error from reading a closed response body, so we
don't need to change any existing ResponseLogHooks that may call close.