more review changes
Make a CheckRetry function type, and move the bulk of the documentation
out of the Client.

Initialize the CheckRetry field to DefaultRetryPolicy in the NewClient
fucntion, and leave out the check in the Do method. Since Logger would
panic immediately, a zero-value Client would never make it to that
check anyway. Also since the default policy was moved from the body of
Do into DefaultRetryPolicy, it makes no sense for it to be nil at all
anyway.